# src/main/resources/api/swagger.yaml
swagger: '2.0'
info:
  version: "1.0.0"
  title: ITAOBAO
basePath: /iTaobao
host: www.ym999.wang:8080
schemes:
  - http
  - https
paths:
...
  /anno:
    get:
      tags: 
        - anno
      description: |
        查询所有公告
      responses:
        200:
          description: Successful response
          schema:
            $ref: '#/definitions/ApiResponse'

    post:
      tags:
        - anno
      description: |
        增加公告
      parameters:
        -
          name: body
          in: body
          description: add anno
          required: true
          schema:
            $ref: '#/definitions/announcement'
      # Expected responses for this operation:
      responses:
        # Response code
        200:
          description: Successful response
          schema:
            $ref: '#/definitions/ApiResponse'
    put:
      tags:
        - anno
      description: |
        修改公告信息
      parameters:
        -
          name: body
          in: body
          description: add anno
          required: true
          schema:
            $ref: '#/definitions/announcement'
      responses:
        # Response code
        200:
          description: Successful response
          schema:
            $ref: '#/definitions/ApiResponse'
    delete:
      tags:
        - anno
      description: | 
        删除公告
      parameters:
        - 
          name: annoID
          in: query
          description: name of user
          required: true
          type: integer
          format: int32
      responses:
        200:
          description: Successful response
          schema:
            $ref: '#/definitions/ApiResponse'
  /business:
    get:
      tags: 
        - bussiness
      # Describe this verb here. Note: you can use markdown
      description: |
        查询所有商户
      responses:
        # Response code
        200:
          description: Successful response
          schema:
            $ref: '#/definitions/ApiResponse'
    post:
      tags:
        - bussiness
      # Describe this verb here. Note: you can use markdown
      description: |
        增加商户
        
      parameters:
        -
          name: body
          in: body
          description: bus
          required: true
          schema:
            $ref: '#/definitions/bussiness'
      # Expected responses for this operation:
      responses:
        # Response code
        200:
          description: Successful response
          schema:
            $ref: '#/definitions/ApiResponse'
    put:
      tags:
        - bussiness
      description: |
      parameters:
        -
          name: body
          in: body
          description: |
            修改商户信息
          required: true
          schema:
            $ref: '#/definitions/bussiness'
      responses:
        # Response code
        200:
          description: Successful response
          schema:
            $ref: '#/definitions/ApiResponse'
    delete:
      tags:
        - bussiness
      description: |
        删除商户
      parameters:
        -
          name: busID
          in: query
          description: bus
          required: true
          type: integer
          format: int32
      responses:
        200:
          description: Successful response
          schema:
            $ref: '#/definitions/ApiResponse'  
  /cate:
    get:
      tags: 
        - cate
      # Describe this verb here. Note: you can use markdown
      description: |
        查询所有库存
        
      responses:
        200:
          description: Successful response
          schema:
            $ref: '#/definitions/ApiResponse'
    post:
      tags:
        - cate
      # Describe this verb here. Note: you can use markdown
      description: |
        增加库存数据
        
      parameters:
        -
          name: body
          in: body
          description: bus
          required: true
          schema:
            $ref: '#/definitions/category'
      responses:
        200:
          description: Successful response
          schema:
            $ref: '#/definitions/ApiResponse'
    put:
      tags:
        - cate
      description: |
        修改库存信息，包括其对应的商品，库存数量
        
      parameters:
        -
          name: body
          in: body
          description: bus
          required: true
          schema:
            $ref: '#/definitions/category'
      responses:
        # Response code
        200:
          description: Successful response
          schema:
            $ref: '#/definitions/ApiResponse'
    delete:
      tags:
        - cate
      description: |
        删除库存数据，单条数据，由传入的ID决定
        
      parameters:
        -
          name: cateID
          in: query
          description: bus
          required: true
          type: integer
          format: int32
      responses:
        200:
          description: Successful response
          schema:
            $ref: '#/definitions/ApiResponse'       
  /coll:
    get:
      tags: 
        - coll
      # Describe this verb here. Note: you can use markdown
      description: |
        查询所有收藏
      # This is array of GET operation parameters:
      responses:
        # Response code
        200:
          description: Successful response
          schema:
            $ref: '#/definitions/ApiResponse'
    post:
      tags:
        - coll
      # Describe this verb here. Note: you can use markdown
      description: |
        添加收藏数据
      # This is array of GET operation parameters:
      parameters:
        -
          name: body
          in: body
          description: bus
          required: true
          schema:
            $ref: '#/definitions/collect'
      responses:
        # Response code
        200:
          description: Successful response
          schema:
            $ref: '#/definitions/ApiResponse'
    put:
      tags:
        - coll
      description: |
        修改收藏信息
        
      parameters:
        -
          name: body
          in: body
          description: bus
          required: true
          schema:
            $ref: '#/definitions/collect'
      responses:
        # Response code
        200:
          description: Successful response
          schema:
            $ref: '#/definitions/ApiResponse'
    delete:
      tags:
        - coll
      description: |
        删除收藏数据，单条数据，由传入的ID决定
        
      parameters:
        - 
          name: collID
          in: query
          description: name of user
          required: true
          type: integer
          format: int32
      responses:
  
        200:
          description: Successful response
          schema:
            $ref: '#/definitions/ApiResponse'      
  /credit:
    get:
      tags: 
        - credit
      # Describe this verb here. Note: you can use markdown
      description: |
        查询所有用户积分信息
      # This is array of GET operation parameters:
      responses:
        # Response code
        200:
          description: Successful response
          schema:
            $ref: '#/definitions/ApiResponse'
    post:
      tags:
        - credit
      # Describe this verb here. Note: you can use markdown
      description: |
        添加积分数据
      # This is array of GET operation parameters:
      parameters:
        -
          name: body
          in: body
          description: bus
          required: true
          schema:
            $ref: '#/definitions/credits'
      responses:
        # Response code
        200:
          description: Successful response
          schema:
            $ref: '#/definitions/ApiResponse'
    put:
      tags:
        - credit
      description: |
        修改积分信息
        
      parameters:
        -
          name: body
          in: body
          description: bus
          required: true
          schema:
            $ref: '#/definitions/credits'
      responses:
        # Response code
        200:
          description: Successful response
          schema:
            $ref: '#/definitions/ApiResponse'
    delete:
      tags:
        - credit
      description: |
        删除积分数据，单条数据，由传入的ID决定
        
      parameters:
        - 
          name: reditsID
          in: query
          description: name of user
          required: true
          type: integer
          format: int32
      responses:
  
        200:
          description: Successful response
          schema:
            $ref: '#/definitions/ApiResponse'       
  /fav:
    get:
      tags: 
        - fav
      # Describe this verb here. Note: you can use markdown
      description: |
        查询所有优惠信息
        
      responses:
        # Response code
        200:
          description: Successful response
          schema:
            $ref: '#/definitions/ApiResponse'
    post:
      tags:
        - fav
      # Describe this verb here. Note: you can use markdown
      description: |
        添加优惠数据
        
      parameters:
        -
          name: body
          in: body
          description: bus
          required: true
          schema:
            $ref: '#/definitions/favourable'
      responses:
        # Response code
        200:
          description: Successful response
          schema:
            $ref: '#/definitions/ApiResponse'
    put:
      tags:
        - fav
      description: |
        修改优惠信息
        
      parameters:
        -
          name: body
          in: body
          description: bus
          required: true
          schema:
            $ref: '#/definitions/favourable'
      responses:
        # Response code
        200:
          description: Successful response
          schema:
            $ref: '#/definitions/ApiResponse'
    delete:
      tags:
        - fav
      description: |
        删除优惠数据，单条数据，由传入的ID决定
        
      parameters:
        - 
          name: userID
          in: query
          description: name of user
          required: true
          type: integer
          format: int32
      responses:
  
        200:
          description: Successful response
          schema:
            $ref: '#/definitions/ApiResponse'     
...
definitions:
...
  announcement:
    type: object
    properties:
      annu_ID:
        type: integer
        format: int32
        description: the ID of annucement
      annu_data:
        type: string
        description: the data of this annucement
  bussiness:
    type: object
    properties:
      bus_ID:
        type: integer
        format: int32
        description: id
      bus_detail_ID:
        type: integer
        format: int32
        description: id of the business'detail
      bus_Name:
        type: string
        description: name of bussiness
      bus_address:
        type: string
        description: address of this bussiness
      bus_phone: 
        type: string
        description: phone number
      bus_email:
        type: string
        description: email of bussiness
      bus_people:
        type: string
        description: l
  category:
    type: object
    properties:
      cate_ID:
        type: integer
        format: int32
        description: d
      pro_ID:
        type: integer
        format: int32
        description: ss
      cate_name:
        type: string
        description: sl
  collect:
    type: object
    properties:
      collectID:
        type: integer
        format: int32
        description: ls 
      userID:
        type: integer
        format: int32
        description: ls
      favourite_status:
        type: integer
        description: se
  credits:
    type: object
    properties:
      creditsID:
        type: integer
        format: int32
        description: ls
      userID:
        type: string
        description: ls
      credits:
        type: integer
        format: int32
        description: dl
  favourable:
    type: object
    properties:
      fav_ID:
        type: integer
        format: int32
        description: ls
      pro_ID:
        type: integer
        format: int32
        description: ls
      fav_name:
        type: string
        description: name of fav
      fav_price:
        type: number
        format: float
        description: l
      fav_method:
        type: integer
        format: int32
        description: l
...
  ApiResponse:
    type: object
    properties:
      falg:
        type: boolean
        description: is sucessful?
      code:
        type: integer
        format: int32
      message:
        type: string
        description: return message